+++ SQuiL.Tests/MutationUnderPlainQueryWarnsSP0023/TransactionDiagnosticTests.MutationUnderPlainQueryWarnsSP0023#MutationUnderPlainQueryWarnsSP0023DataContext.g.verified.cs
@@ -28,7 +28,7 @@ partial class MutationUnderPlainQueryWarnsSP0023DataContext : SQuiLBaseDataConte
 		
 		List<DbParameter> parameters = new()
 		{
-			CreateParameter("@Param_Id", System.Data.SqlDbType.BigInt, request.Id)
+			CreateParameter("@Param_Id", System.Data.SqlDbType.Int, request.Id)
 		};
 		
 		command.CommandText = Query(parameters);
