+++ SQuiL.Tests/OwnBeginTranUnderEnabledTransactionErrorsSP0025/TransactionDiagnosticTests.OwnBeginTranUnderEnabledTransactionErrorsSP0025#OwnBeginTranUnderEnabledTransactionErrorsSP0025DataContext.g.verified.cs
@@ -28,7 +28,7 @@ partial class OwnBeginTranUnderEnabledTransactionErrorsSP0025DataContext : SQuiL
 		
 		List<DbParameter> parameters = new()
 		{
-			CreateParameter("@Param_Id", System.Data.SqlDbType.BigInt, request.Id)
+			CreateParameter("@Param_Id", System.Data.SqlDbType.Int, request.Id)
 		};
 		
 		command.CommandText = Query(parameters);
